Renamed proxy helper package to reflect it is not just for konnectivity
